+++ src/main/java/com/community/petish/mypage/mapper/MessageMapper.java
@@ -12,25 +12,25 @@ public interface MessageMapper {
 	public int writeMessage(MessageRequestDTO dto);
 	
 	//2. 받은 쪽지 리스트 조회
-	public ArrayList<MessageResponseDTO> getReceivedMessageList(int user_id);
+	public ArrayList<MessageResponseDTO> getReceivedMessageList(Long user_id);
 	
 	//3. 보낸 쪽지 리스트 조회
-	public ArrayList<MessageResponseDTO> getSentMessageList(int user_id);
+	public ArrayList<MessageResponseDTO> getSentMessageList(Long user_id);
 	
 	//4. 받은 쪽지 상세조회
-	public MessageResponseDTO receivedMessageDetail(int id);
+	public MessageResponseDTO receivedMessageDetail(Long id);
 	
 	//5. 보낸 쪽지 상세조회
-	public MessageResponseDTO sentMessageDetail(int id);
+	public MessageResponseDTO sentMessageDetail(Long id);
 	
 	//6. 쪽지 삭제
-	public int deleteMessage(int id);
+	public int deleteMessage(Long id);
 	
 	//7. 받은 메세지중 삭제안된 메세지 갯수
-	public int getUndeletedReceived(int user_id);
+	public int getUndeletedReceived(Long user_id);
 	
 	//8. 보낸 메세지중 삭제안된 메세지 갯수
-	public int getUndeletedSent(int user_id);
+	public int getUndeletedSent(Long user_id);
 	
 	//9. 쪽지 읽음으로 변경
 	public int changeReadAttr(MessageRequestDTO dto);
